[OWL-1695] Move tests & rename file.
